moved inmem stores into dedicated package
